requirements.txt
# Python version 3.9.6
#
# These requirements are used for local development, not for the various scripts which should have their own
# requirement files for the build processes
python-dotenv==0.19.0
google-cloud-storage==1.41.1
google_cloud_secret_manager==2.7.0
flask==2.0.1
